Refactor procdump/profiler
