+++ include/Includes.h
@@ -1,3 +1,4 @@
+#include "ProfilerCommon.h"
 #include "CoreDumpWriter.h"
 #include "Events.h"
 #include "GenHelpers.h"
@@ -9,3 +10,4 @@
 #include "Process.h"
 #include "DotnetHelpers.h"
 #include "ProfilerHelpers.h"
+
